ADM-808 refactor
