--- frontend/e2e/fixtures/consts.ts
@@ -1,1 +1,0 @@
-export const viewportDefault = { width: 1920, height: 1080 };
