+++ frontend/e2e/fixtures/index.ts
@@ -1,3 +1,5 @@
+export const VIEWPORT_DEFAULT = { width: 1920, height: 1080 };
+
 export const E2E_PROJECT_NAME = 'E2EProjectName';
 
 export const CONFIG_STEP_SAVING_FILENAME = 'config-step.json';
